Support config file re-definition of the active click regions.
To support languages that might use longer words than English for the
action regions (New, Solve, Quit etc) we need to read the regions from
a config file. This will also permit using backgrounds that move everything
around if it is extended to define the positions of the text boxes and
clock as well.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>